+++ packages/stacked/lib/src/view_model_builder.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'base_view_models.dart';
@@ -93,22 +92,9 @@ class _ViewModelBuilderState<T extends ChangeNotifier>
 
   void _initialiseSpecialViewModels() {
     // Add any additional actions here for spcialised ViewModels
-    // TODO: Provide a closed implemenation of this functionality. Refer to the Open Closed
-    // principle in the SOLID principles
-    if (_model is FutureViewModel) {
-      (_model as FutureViewModel).runFuture();
-    }
-
-    if (_model is MultipleFutureViewModel) {
-      (_model as MultipleFutureViewModel).runFutures();
-    }
-
-    if (_model is StreamViewModel) {
-      (_model as StreamViewModel).initialise();
-    }
-
-    if (_model is MultipleStreamViewModel) {
-      (_model as MultipleStreamViewModel).initialise();
+     // Add any additional actions here for spcialised ViewModels
+    if (_model is Initialisable) {
+      (_model as Initialisable).initialise();
     }
   }
 
@@ -158,5 +144,5 @@ class _ViewModelBuilderState<T extends ChangeNotifier>
   }
 }
 
-/// EXPERIMENTAL: Returns the ViewModel provided above this widget in the tree 
+/// EXPERIMENTAL: Returns the ViewModel provided above this widget in the tree
 T getParentViewModel<T>(BuildContext context) => Provider.of<T>(context);