+++ packages/stacked_services/lib/src/navigation_service.dart
@@ -12,6 +12,7 @@ class NavigationTransition {
   static const String Rotate = 'rotate';
   static const String Size = 'size';
   static const String RightToLeftWithFade = 'righttoleftwithfade';
+  static const String LeftToRighttWithFade = 'lefttorightwithfade';
   static const String Cupertino = 'cupertino';
 }
 
@@ -30,16 +31,18 @@ class NavigationService {
     NavigationTransition.Rotate: Transition.rotate,
     NavigationTransition.Size: Transition.size,
     NavigationTransition.RightToLeftWithFade: Transition.rightToLeftWithFade,
-    NavigationTransition.RightToLeftWithFade: Transition.leftToRightWithFade,
+    NavigationTransition.LeftToRighttWithFade: Transition.leftToRightWithFade,
     NavigationTransition.Cupertino: Transition.cupertino,
   };
 
-  get navigatorKey => Get.key;
+  get navigatorKey {
+    return Get.key;
+  }
 
   /// Allows you to configure the default behaviour for navigation.
   ///
   /// [defaultTransition] can be set using the static members of [NavigationTransition]
-  /// 
+  ///
   /// If you want to use the string directly. Defined [transition] values are
   /// - fade
   /// - rightToLeft
@@ -72,7 +75,7 @@ class NavigationService {
   /// of routeName (String).
   ///
   /// Defined [transition] values can be accessed as static memebers of [NavigationTransition]
-  /// 
+  ///
   /// If you want to use the string directly. Defined [transition] values are
   /// - fade
   /// - rightToLeft
@@ -132,6 +135,11 @@ class NavigationService {
     return Get.toNamed(routeName, arguments: arguments);
   }
 
+  /// Pushes [view] onto the navigation stack
+  Future<dynamic> navigateToView(Widget view, {dynamic arguments}) {
+    return Get.to(view, arguments: arguments);
+  }
+
   /// Replaces the current route with the [routeName]
   Future<dynamic> replaceWith(String routeName, {dynamic arguments}) {
     return Get.offNamed(routeName, arguments: arguments);
@@ -139,9 +147,7 @@ class NavigationService {
 
   /// Clears the entire back stack and shows [routeName]
   Future<dynamic> clearStackAndShow(String routeName, {dynamic arguments}) {
-    _clearBackstackCompletely();
-
-    return replaceWith(routeName, arguments: arguments);
+    return Get.offAllNamed(routeName);
   }
 
   /// Pops the navigation stack until there's 1 view left then pushes [routeName] onto the stack
@@ -151,6 +157,13 @@ class NavigationService {
     return navigateTo(routeName, arguments: arguments);
   }
 
+  /// Pops the navigation stack until there's 1 view left then pushes [view] onto the stack
+  Future<dynamic> clearTillFirstAndShowView(Widget view, {dynamic arguments}) {
+    _clearBackstackTillFirst();
+
+    return navigateToView(view);
+  }
+
   /// Push route and clear stack until predicate is satisfied
   Future<dynamic> pushNamedAndRemoveUntil(String routeName,
       {RoutePredicate predicate, arguments, int id}) {
@@ -158,12 +171,8 @@ class NavigationService {
         predicate: predicate, arguments: arguments, id: id);
   }
 
-  void _clearBackstackCompletely() {
-    navigatorKey.currentState.popUntil((route) => false);
-  }
-
   void _clearBackstackTillFirst() {
-    navigatorKey.currentState.popUntil((route) => route.isFirst);
+    navigatorKey.currentState.popUntil((Route route) => route.isFirst);
   }
 
   Transition _getTransitionOrDefault(String transition) {