+++ packages/stacked_services/lib/stacked_services.dart
@@ -1,5 +1,5 @@
 library stacked_services;
 
 export 'src/navigation_service.dart';
-export 'src/dialog_service.dart';
+export 'src/dialog/dialog_service.dart';
 export 'src/snackbar_service.dart';